__mock_device__/requirement.txt
Werkzeug==2.2.2
paho-mqtt==1.6.1